Add unit test for checkout cancellation
Summary: In this diff, add two unit tests capturing the behavior of cancelling checkout and leaving the repo in an interrupted state. This state can be recovered from with another checkout. The two unit tests show that checkout can be cancelled in two different points in the flow, and still recover.

Reviewed By: muirdm

Differential Revision: D87754738

fbshipit-source-id: c2751063b0038ba8abae8dd4f9b26d012a4c8649