Increase SIGTERM time to shutdown to 20 seconds
Summary:
We are adding cancellation to EdenFS, and we want to be able to give `checkout` the most time to complete before cancelling and potentially entering an interrupted checkout state.

Previous to this diff, `checkout` would have up to 15 seconds to complete before `_exit(1)`.

After this diff,  `checkout` would have up to 15 seconds to complete before triggering cancellation, and 5 remaining seconds to wrap up after cancellation. Leaving a total of 20 seconds before `_exit(1)`.

Reviewed By: MichaelCuevas

Differential Revision: D88869686

fbshipit-source-id: d0ce12d36a84951fbdf3257c755129f4928d7c33