Instrument checkout with cancellation
Summary:
In order to support cancelling 'checkout' Eden Thrift requests: we need to first need to plumb the cancellation token through the call stack, we second need to cooperatively check the token.

The typical usage is with a `CheckoutContext` by: `ctx->throwIfCanceled();` -- This will check the Eden Config before throwing, allowing us to slowroll this feature.

In this initial instrumentation I've picked the following 5 points during checkout:

- Inside the `computeDiff` loop: TreeInode.cpp:3104
- Before starting the recursive tree checkout: EdenMount.cpp:1617
- Inside the 'checkout' directory recursive loop: TreeInode.cpp:3248
- At each checkoutAction, for each modified file/dir: CheckoutAction.cpp:67
- Before journal diff: EdenMount.cpp:1543

Reviewed By: muirdm

Differential Revision: D87752351

fbshipit-source-id: a5ee3fffc490b5676eb0d82bd8aa38aca808c1da