http-client: improve unix socket errors
Summary:
Annotate curl errors with the underlying os_error when available. This gives better error context when there is an error connecting to a unix domain socket.

There are various code paths that propagate curl errors, so it was a bit annoying. I moved the unix domain socket onto the RequestContext so we can get it from the easy handle (to include in the error message).

If, for example, the unix domain socket doesn't exist, we would previously see an error like:

    Could not connect to server (Failed to connect to <HOSTNAME> port 80 after 0 ms: Could not connect to server

This was confusing because we didn't actually perform an HTTP request. Now it shows up like:

    Could not connect to server (Failed to connect to <HOSTNAME> port 80 after 0 ms: Could not connect to server (<SOCKET PATH>: No such file or directory (os error 2)))

Reviewed By: MichaelCuevas

Differential Revision: D89490357

fbshipit-source-id: e0bf7e19c62183a53cb7070cbe9895595c60908d