repos_manager: reduce concurrency from 50 to 30 also for updates
Summary:
D89373726 changed to buffer_unordered and increased concurrency to 50.
We believe this could be causing the CPU spikes that are causing the health check failures in bookmark service: https://fburl.com/services/tgexjjbs.

D89469378 changed the concurrency value for the population, but didn't include the updates. Since we are still reaching 97% of CPU, I'm reducing the updates to see if this helps.

Reviewed By: chiara125

Differential Revision: D89537599

fbshipit-source-id: 9dceff7e49deecba285e5086c5e73e34324b95c1