Run jf unlink when Grafting Public Commits to FRBranch
Summary:
When we're doing a grafting operation to a full-repo branch, oftentimes we're grafting a public commit (from trunk or a different branch) to this one. We can use the Download Commits UI in ISL to run `sl graft`, but to actually commit it to the branch we have to manually run `jf unlink` and then resubmit as a new Diff first.

This has appeared as questions in other branches like Pick branches before (https://fb.workplace.com/groups/scm/permalink/5998876783495225/)

In this diff we update the Download Commits behavior to check if it's 1) a closed commit, and 2) rebasing on a full-repo branch — in this case we will run `jf unlink` automatically after grafting. I can't imagine a case where we _wouldn't_ want to run `jf unlink` after grafting, since downloading a closed commit isn't very useful otherwise

Differential Revision: D88337798

fbshipit-source-id: 13921eb187ebf6c76b3ceb0264cb4ebfac335824