pull: support pull fast path if old "master" is not present
Summary:
Previously, the fast path pull can only run when the old "master" commit exists
locally. This was a limitation of the `pull_ff_master` API taking two explicit
commits, instead of hashes.

We had long migrated to the more general purposed `fastpull` API that takes
sets. So let's just update related fast path to support pulling from an empty
repo.

This is intended to be used in a shallow repo cloned from a virtual repo.

Reviewed By: muirdm

Differential Revision: D89238753

fbshipit-source-id: 6fff7df282ed7930f7d929f879d7996cc7673239